refactor: move makeAura to utils.js
